Coerce casts what it can and not all-or-none
utilities._coerce used to take in an object and try to change the
entire object's dtype. Now, it changes only what columns can be
changeable.

For example, if there's a column of words and column of integers,
coerce cam now cast only the column of integers to type integer
and leave the column of words as pandas objects.

Add unit tests for coerce.