Redirect mapservice to census for ACS 2020
Catch and redirect if ACS 2020 to Census 2020. Could be done in tiger, but this way it will be assigned to attribute. Could be done earlier in product but suspect this will be more concise. Could also be done as a try/catch instead, but current key error handling is good.